requirements.txt
fastapi
langchain
langchain-tavily
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-google-genai
langchain-pinecone
pinecone-client
langsmith
numpy
openai
pydantic
pydantic_core
pypdf
python-dotenv
PyJWT
slowapi
python-multipart
redis
langchain-redis
tavily-python
uvicorn[standard]
pymupdf4llm
